+++ cynic-parser/benches/parsing-benchmark.rs
@@ -1,5 +1,5 @@
 use criterion::{black_box, criterion_group, criterion_main, Criterion};
-use cynic_parser::AstBuilder;
+use cynic_parser::Ast;
 
 pub fn criterion_benchmark(c: &mut Criterion) {
     let input = "type MyType { field: Whatever, field: Whatever }";
@@ -7,7 +7,7 @@ pub fn criterion_benchmark(c: &mut Criterion) {
         b.iter(|| {
             let lexer = cynic_parser::Lexer::new(input);
             let object = cynic_parser::ObjectParser::new()
-                .parse(input, &mut AstBuilder::new(), lexer)
+                .parse(input, &mut Ast::new(), lexer)
                 .unwrap();
             black_box(object)
         })
