+++ cynic-parser/src/schema.lalrpop
@@ -1,14 +1,13 @@
 use std::str::FromStr;
 
-use crate::ast::{Schema, Field, Object};
 use crate::lexer;
 
-use crate::builder::{
-    RootOperationTypeDefinition, OperationType, AstBuilder, NodeId, SchemaDefinition,
+use crate::ast::{
+    RootOperationTypeDefinition, OperationType, Ast, NodeId, SchemaDefinition,
     ObjectDefinition, FieldDefinition
 };
 
-grammar<'input>(input: &'input str, builder: &mut AstBuilder);
+grammar<'input>(input: &'input str, ast: &mut Ast);
 
 pub SchemaDefinition: SchemaDefinition = {
     schema "{" <roots:RootOperationTypeDefinition*> "}" => SchemaDefinition { <> }
@@ -30,11 +29,11 @@ FieldDefinitions: Vec<NodeId> = {
 };
 
 FieldDefinition: NodeId = <name:Name> ":" <ty:NamedType> =>
-    builder.field_definition(FieldDefinition { <> });
+    ast.field_definition(FieldDefinition { <> });
 
-Name: NodeId = <s:Ident> => builder.ident(s);
+Name: NodeId = <s:Ident> => ast.ident(s);
 
-NamedType: NodeId = <s:Ident> => builder.ident(s);
+NamedType: NodeId = <s:Ident> => ast.ident(s);
 
 // TODO: Make this NodeId probably...
 Ident: &'input str = {
